Refactor dashboard shell into modular components
